Fix _xgetbv return type to long long to match GCC 13 declaration

+++ legacy/trunk/util/m_misc.cpp
@@ -46,10 +46,10 @@
 // GCC 13+ / MinGW-w64 _xgetbv conflict workaround
 #ifdef __MINGW32__
 #undef _xgetbv
-static inline unsigned long long _xgetbv(unsigned int index) {
-    unsigned long long eax, edx;
+static inline long long _xgetbv(unsigned int index) {
+    unsigned int eax, edx;
     __asm__ __volatile__ ("xgetbv" : "=a"(eax), "=d"(edx) : "c"(index));
-    return (edx << 32) | eax;
+    return ((long long)edx << 32) | eax;
 }
 #endif
